merged

+++ app/Http/Controllers/Expense/ExpenseOutgoingController.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Expense;
 
+use App\Actions\Sempoa\PushToJurnalSempoa;
+use App\Http\Controllers\Controller;
 use App\Models\Expense;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
@@ -26,39 +28,23 @@ class ExpenseOutgoingController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
      * Display the specified resource.
      */
     public function show(Expense $expense_outgoing)
     {
-        $data = [
-            'title' => "Realisasi Pengeluaran Biaya",
-            'expense' => $expense_outgoing,
-        ];
-        return view('pages.expense-outgoing.detail', $data);
-    }
+        try {
+            if ($expense_outgoing->status != Expense::STATUS_APPROVED)
+                throw new \Exception('Transaksi tidak bisa dilakukan. Status harus "Disetujui".');
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
+            $data = [
+                'title' => "Realisasi Pengeluaran Biaya",
+                'expense' => $expense_outgoing,
+            ];
+        } catch (\Throwable $th) {
+            return redirect()->back()->withToastError('Ops! ' . $th->getMessage());
+        }
+
+        return view('pages.expense-outgoing.detail', $data);
     }
 
     /**
@@ -67,21 +53,22 @@ class ExpenseOutgoingController extends Controller
     public function update(Request $request, Expense $expense_outgoing)
     {
         DB::beginTransaction();
-        
+
         try {
+            if ($expense_outgoing->status != Expense::STATUS_APPROVED)
+                throw new \Exception('Transaksi tidak bisa dilakukan. Status harus "Disetujui".');
+
             $expense_outgoing->approval_by = Auth::user()->id;
             $expense_outgoing->status = Expense::STATUS_DONE;
             $expense_outgoing->expense_outgoing_date = now();
-              
-            // Upload Expense Photo 
-              if ($request->hasFile('file_photo')) {
+
+            // Upload Expense Photo
+            if ($request->hasFile('file_photo')) {
                 $uploadedFile = $request->file('file_photo');
                 if ($expense_outgoing->file_photo) Storage::delete($expense_outgoing->getRawOriginal('file_photo')); // Delete old photo
                 $expense_outgoing->file_photo = Storage::putFileAs('expense_photo', $uploadedFile, $uploadedFile->hashName());
-            } 
+            }
             $expense_outgoing->save();
-
-            DB::commit();
 
             // pengurangan saldo jika diterima
             if ($request->action == 'approve') {
@@ -90,21 +77,16 @@ class ExpenseOutgoingController extends Controller
                 }
             }
 
+            DB::commit();
+
+            PushToJurnalSempoa::handle($expense_outgoing);
             // Notification
             $expense_outgoing->requested_by->notify(new ExpenseOutgoingNotification($expense_outgoing));
-
-            return redirect()->route('expense-outgoing.index')->withToastSuccess('Berhasil mengubah Status!');
         } catch (\Throwable $th) {
             DB::rollBack();
-            return redirect()->back()->withToastError('Ops, ada kesalahan saat mengubah Status!');
+            return redirect()->back()->withToastError('Ops! ' . $th->getMessage());
         }
-    }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-        //
+        return redirect()->route('expense-outgoing.index')->withToastSuccess('Berhasil mengubah Status!');
     }
 }
